pyproject.toml
[project]
name = "agent-run-optimizer"
version = "0.1.0"
description = "Make LLM agent executions observable, replayable, and analyzable via graph-based run tracking"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "networkx>=3.3",
    "rich>=13.0",
    "pydantic>=2.0",
    "click>=8.1",
    "pyyaml>=6.0",
    "filelock>=3.13",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-api>=1.25",
]

[dependency-groups]
dev = [
    "black>=24.0",
    "flake8>=7.0",
    "flake8-docstrings>=1.7",
    "isort>=5.13",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_run_optimizer"]

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
exclude = ["*.ipynb", "*__pycache__*", "*test*"]
lint.select = ["RUF022"]

[tool.pytest.ini_options]
python_files = "tests.py test_*.py *_tests.py"
asyncio_mode = "auto"